+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - May 11, 2025
+
+**First update of 2025!**
+
+### Added
+
+- Add the Goodreads Choice Awards route - [(ISSUE)](https://github.com/nesaku/BiblioReads/issues/30)
+
+### Changed
+
+- Make the CoverImage loading priority configurable
+
+### Fixed
+
+- Fixed the library page background not reaching the bottom of the page when more than three books are added
+
 ## [3.0.0] (Holo) - Dec 26, 2024
 
 ### Updated to Next.js 15 - Possible Breaking Changes